Rename and restructure section types
